fix: move memory stats to ContactInfo Job
